requirements.txt
torch==0.4.1
torchvision==0.2.1
visdom==0.1.8.3
imageio==2.5.0
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.9.0=pypi_0
ca-certificates=2020.1.1=0
cachetools=4.1.0=pypi_0
certifi=2019.11.28=py36_0
chardet=3.0.4=pypi_0
cvbase=0.5.5=pypi_0
cycler=0.10.0=pypi_0
decorator=4.4.2=pypi_0
google-auth=1.15.0=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
grpcio=1.29.0=pypi_0
idna=2.9=pypi_0
imageio=2.5.0=pypi_0
imageio-ffmpeg=0.4.1=pypi_0
importlib-metadata=1.6.0=pypi_0
kiwisolver=1.1.0=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
markdown=3.2.2=pypi_0
matplotlib=3.2.0=pypi_0
ncurses=6.2=he6710b0_0
networkx=2.4=pypi_0
numpy=1.18.1=pypi_0
oauthlib=3.1.0=pypi_0
opencv-python=4.2.0.32=pypi_0
openssl=1.1.1d=h7b6447c_4
pillow=6.0.0=pypi_0
pip=20.0.2=py36_1
protobuf=3.12.1=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyparsing=2.4.6=pypi_0
python=3.6.10=h0371630_0
python-dateutil=2.8.1=pypi_0
pywavelets=1.1.1=pypi_0
pyyaml=5.3=pypi_0
pyzmq=19.0.0=pypi_0
readline=7.0=h7b6447c_5
requests=2.23.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.0=pypi_0
scikit-image=0.16.2=pypi_0
scipy=1.4.1=pypi_0
setuptools=45.2.0=py36_0
six=1.14.0=pypi_0
sqlite=3.31.1=h7b6447c_0
tensorboard=2.2.1=pypi_0
tensorboard-plugin-wit=1.6.0.post3=pypi_0
terminaltables=3.1.0=pypi_0
tk=8.6.8=hbc83047_0
torch=0.4.1=pypi_0
torchfile=0.1.0=pypi_0
torchvision=0.2.1=pypi_0
tornado=6.0.3=pypi_0
urllib3=1.25.8=pypi_0
visdom=0.1.8.3=pypi_0
websocket-client=0.57.0=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.34.2=py36_0
xz=5.2.4=h14c3975_4
zipp=3.1.0=pypi_0
zlib=1.2.11=h7b6447c_3
